requirements.txt
streamlit
openai==1.58.1
chromadb==1.5.0
python-dotenv
pandas
lxml==5.2.1
matplotlib==3.9.2
tqdm